streamlit/requirements.txt
pandas~=2.1.1
numpy~=1.26.1
matplotlib~=3.8.0
streamlit-awesome-table
